only create icons on main branch

+++ install.sh
@@ -1628,30 +1628,33 @@ fi
 # We mark the script as finished	
 echo "" > ~/emudeck/.finished
 
-#We create new icons
-rm -rf ~/Desktop/EmuDeckUninstall.desktop &>> /dev/null
-echo '#!/usr/bin/env xdg-open
-[Desktop Entry]
-Name=Uninstall EmuDeck
-Exec=curl https://raw.githubusercontent.com/dragoonDorise/EmuDeck/main/uninstall.sh | bash -s -- SD
-Icon=delete
-Terminal=true
-Type=Application
-StartupNotify=false' > ~/Desktop/EmuDeckUninstall.desktop
-chmod +x ~/Desktop/EmuDeckUninstall.desktop
+if [ $branch == 'main' ];then
 
-rm -rf ~/Desktop/EmuDeck.desktop &>> /dev/null
-rm -rf ~/Desktop/EmuDeckSD.desktop &>> /dev/null
-echo "#!/usr/bin/env xdg-open
-[Desktop Entry]
-Name=EmuDeck (${version})
-Exec=curl https://raw.githubusercontent.com/dragoonDorise/EmuDeck/main/install.sh | bash -s -- SD
-Icon=steamdeck-gaming-return
-Terminal=true
-Type=Application
-StartupNotify=false" > ~/Desktop/EmuDeck.desktop
-chmod +x ~/Desktop/EmuDeck.desktop
+	#We create new icons
+	rm -rf ~/Desktop/EmuDeckUninstall.desktop &>> /dev/null
+	echo '#!/usr/bin/env xdg-open
+	[Desktop Entry]
+	Name=Uninstall EmuDeck
+	Exec=curl https://raw.githubusercontent.com/dragoonDorise/EmuDeck/main/uninstall.sh | bash -s -- SD
+	Icon=delete
+	Terminal=true
+	Type=Application
+	StartupNotify=false' > ~/Desktop/EmuDeckUninstall.desktop
+	chmod +x ~/Desktop/EmuDeckUninstall.desktop
+	
+	rm -rf ~/Desktop/EmuDeck.desktop &>> /dev/null
+	rm -rf ~/Desktop/EmuDeckSD.desktop &>> /dev/null
+	echo "#!/usr/bin/env xdg-open
+	[Desktop Entry]
+	Name=EmuDeck (${version})
+	Exec=curl https://raw.githubusercontent.com/dragoonDorise/EmuDeck/main/install.sh | bash -s -- SD
+	Icon=steamdeck-gaming-return
+	Terminal=true
+	Type=Application
+	StartupNotify=false" > ~/Desktop/EmuDeck.desktop
+	chmod +x ~/Desktop/EmuDeck.desktop
 
+fi
 
 echo -e "Cleaning up downloaded files..."	
 rm -rf ~/dragoonDoriseTools
